LangChain Pipeline/requirements.txt
langchain
langchain-openai
langchain-community
wikipedia
python-dotenv